Fix for an apparent bug introduced in Firefox 56.0:
As of Firefox 56.0, the contents of an SVG element no longer renders
if the "transform-style" property of its immediate parent element is
"preserve-3d". So the icons on the block faces are not rendering in
Firefox 56.0 since each block face is an SVG element that's a child of
a div of class "block" which sets "transform-style" to "preserve-3d"!

Implemented a workaround: wrap each SVG element corresponding to a
block face in another div. That div now has class "blockface" and is
a child of a div of class "block". The SVG element no longer has its
class attribute set.